refactor(backoffice-v2): removed alert-definitions domain and moved logic to alerts domain
