--- apps/backoffice-v2/src/domains/alert-definitions/query-keys.ts
@@ -1,15 +1,0 @@
-import { createQueryKeys } from '@lukemorales/query-key-factory';
-import { fetchAlertDefinitionByAlertId } from '@/domains/alert-definitions/fetchers';
-
-export const alertDefinitionsQueryKeys = createQueryKeys('alerts', {
-  byAlertId: ({ alertId }: { alertId: string }) => {
-    return {
-      queryKey: [
-        {
-          alertId,
-        },
-      ],
-      queryFn: () => fetchAlertDefinitionByAlertId({ alertId }),
-    };
-  },
-});
