+++ apps/backoffice-v2/src/domains/alerts/fetchers.ts
@@ -172,3 +172,18 @@ export const updateAlertsDecisionByIds = async ({
 
   return handleZodError(error, alerts);
 };
+
+export const AlertDefinitionByAlertIdSchema = z.object({
+  name: z.string(),
+  description: z.string(),
+});
+
+export const fetchAlertDefinitionByAlertId = async ({ alertId }: { alertId: string }) => {
+  const [alertDefinition, error] = await apiClient({
+    url: `${getOriginUrl(env.VITE_API_URL)}/api/v1/external/alerts/${alertId}/alert-definition`,
+    method: Method.GET,
+    schema: AlertDefinitionByAlertIdSchema,
+  });
+
+  return handleZodError(error, alertDefinition);
+};
