+++ apps/backoffice-v2/src/domains/alerts/hooks/queries/useAlertDefinitionByAlertIdQuery/useAlertDefinitionByAlertIdQuery.tsx
@@ -1,12 +1,12 @@
 import { useIsAuthenticated } from '@/domains/auth/context/AuthProvider/hooks/useIsAuthenticated/useIsAuthenticated';
 import { useQuery } from '@tanstack/react-query';
-import { alertDefinitionsQueryKeys } from '@/domains/alert-definitions/query-keys';
+import { alertsQueryKeys } from '@/domains/alerts/query-keys';
 
 export const useAlertDefinitionByAlertIdQuery = ({ alertId }: { alertId: string }) => {
   const isAuthenticated = useIsAuthenticated();
 
   return useQuery({
-    ...alertDefinitionsQueryKeys.byAlertId({ alertId }),
+    ...alertsQueryKeys.alertDefinitionByAlertId({ alertId }),
     enabled: isAuthenticated && !!alertId,
     staleTime: 100_000,
   });
